update ext subsc

+++ everdues/contracts/root/EverduesRoot.sol
@@ -105,8 +105,8 @@ contract EverduesRoot is EverduesRootBase {
 				mapping(address => ServiceDeployParams),
 				address,
 				//mapping(uint8 => mapping(uint256 => ExternalSubscription)),
-				mapping(uint8 => string),
-				mapping (uint8=>string[]),
+				mapping(uint32 => string),
+				mapping (uint32=>string[]),
 				address// TODO: add ,address (watcher)
 			)
 		);
